+++ lib/mod/restore.sh
@@ -156,8 +156,8 @@ initialize_duplicacy() {
   if [ ! -d "${LOCAL_DIR}" ]; then
     mkdir -p -m 0755 "${LOCAL_DIR}"
     # If variables available, use them for ownership
-    if [ -n "$INVOKING_UID" ] && [ -n "$INVOKING_GID" ]; then
-      chown "$INVOKING_UID":"$INVOKING_GID" "${LOCAL_DIR}"
+    if [ -n "${INVOKING_UID}" ] && [ -n "${INVOKING_GID}" ]; then
+      chown "${INVOKING_UID}":"${INVOKING_GID}" "${LOCAL_DIR}"
     fi
   fi
   
@@ -291,6 +291,10 @@ choose_revision() {
 
 restore_repository() {
   duplicacy restore -r "${REVISION}" -key "${DUPLICACY_RSA_PRIVATE_KEY_FILE}"
+  # Fix ownership of restored files to match the invoking user
+  if [ -n "${INVOKING_UID}" ] && [ -n "${INVOKING_GID}" ]; then
+    chown -R "${INVOKING_UID}":"${INVOKING_GID}" "${LOCAL_DIR}"
+  fi
   echo "Repository restored."
 }